chore: move specs into test/spec folder
