+++ Gruntfile.js
@@ -57,8 +57,8 @@ module.exports = function( grunt ) {
 		watch: {
 			test: {
 				files: [
-					'test/*.spec.js',
-					'test/page/*.js'
+					'test/spec/*.spec.js',
+					'test/page/*.page.js'
 				],
 				tasks: [ 'eslint:test', 'protractor:localhost' ]
 			},
